* Makefile.am (proxycheck): Add new test target. * Makefile.in: Regenerated.
* src/Makefile.am (proxycheck): New test target.
* src/sanity.sh: Keep one more level of check.log backup.
* src/Makefile.in: Regenerated.

CVS-ID: src/ChangeLog 1.2986
CVS-ID: src/Makefile.am 1.43
CVS-ID: src/Makefile.in 1.130
CVS-ID: src/sanity.sh 1.988
CVS-ID: ChangeLog 1.1114
CVS-ID: Makefile.am 1.21
CVS-ID: Makefile.in 1.147

+++ src/sanity.sh
@@ -543,9 +543,8 @@ LOGFILE=`pwd`/check.log
 # Save the previous log in case the person running the tests decides
 # they want to look at it.  The extension ".plog" is chosen for consistency
 # with dejagnu.
-if test -f check.log; then
-	mv check.log check.plog
-fi
+test -f check.plog && mv check.plog check.plog~
+test -f check.log && mv check.log check.plog
 
 # Create the log file so check.log can be tailed almost immediately after
 # this script is started.  Otherwise it can take up to a minute or two before
